Remove lifetimes from Error enums
Although it is possible to return an enum Error with a lifetime,
it very quickly becomes annoying, because unwrapping with ? may
reference local variables.

As a compromise, use a fixed length string, that in the worst
case will store a trimmed string. This should give enough context
to find out what the value is.